+++ SpendiQ/app/src/main/java/com/isis3510/spendiq/viewmodel/AccountViewModel.kt
@@ -104,34 +104,6 @@ class AccountViewModel : ViewModel() {
         }
     }
 
-    class TransactionViewModel(private val repository: AccountRepository = AccountRepository()) : ViewModel() {
-        private val _transactions = MutableStateFlow<List<Transaction>>(emptyList())
-        val transactions: StateFlow<List<Transaction>> = _transactions
-
-        private val _uiState = MutableStateFlow<UiState>(UiState.Idle)
-        val uiState: StateFlow<UiState> = _uiState
-
-        fun fetchTransactions(accountName: String) {
-            viewModelScope.launch {
-                _uiState.value = UiState.Loading
-                try {
-                    val result = repository.fetchTransactions_repo(accountName)
-                    _transactions.value = result
-                    _uiState.value = UiState.Success
-                } catch (e: Exception) {
-                    _uiState.value = UiState.Error(e.message ?: "Unknown error occurred")
-                }
-            }
-        }
-
-        sealed class UiState {
-            data object Idle : UiState()
-            data object Loading : UiState()
-            data object Success : UiState()
-            data class Error(val message: String) : UiState()
-        }
-    }
-
     fun addTransactionWithAccountCheck(transaction: Transaction) {
         viewModelScope.launch {
             _uiState.value = UiState.Loading
