+++ SpendiQ/app/src/main/java/com/isis3510/spendiq/viewmodel/TransactionViewModel.kt
@@ -1,0 +1,37 @@
+package com.isis3510.spendiq.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.isis3510.spendiq.model.data.Transaction
+import com.isis3510.spendiq.model.repository.AccountRepository
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+
+class TransactionViewModel(private val repository: AccountRepository = AccountRepository()) : ViewModel() {
+    private val _transactions = MutableStateFlow<List<Transaction>>(emptyList())
+    val transactions: StateFlow<List<Transaction>> = _transactions
+
+    private val _uiState = MutableStateFlow<UiState>(UiState.Idle)
+    val uiState: StateFlow<UiState> = _uiState
+
+    fun fetchTransactions(accountName: String) {
+        viewModelScope.launch {
+            _uiState.value = UiState.Loading
+            try {
+                val result = repository.fetchTransactions_repo(accountName)
+                _transactions.value = result
+                _uiState.value = UiState.Success
+            } catch (e: Exception) {
+                _uiState.value = UiState.Error(e.message ?: "Unknown error occurred")
+            }
+        }
+    }
+
+    sealed class UiState {
+        data object Idle : UiState()
+        data object Loading : UiState()
+        data object Success : UiState()
+        data class Error(val message: String) : UiState()
+    }
+}
