+++ SpendiQ/app/src/main/java/com/isis3510/spendiq/views/accounts/AccountTransactionsScreen.kt
@@ -27,7 +27,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.google.firebase.Timestamp
 import com.isis3510.spendiq.model.data.Transaction
-import com.isis3510.spendiq.viewmodel.AccountViewModel.TransactionViewModel
+import com.isis3510.spendiq.viewmodel.TransactionViewModel
 import java.text.NumberFormat
 import java.text.SimpleDateFormat
 import java.util.*
